% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{apply_duration_thresh}
\alias{apply_duration_thresh}
\title{Apply a drought duration threshold to a time series}
\usage{
apply_duration_thresh(tseries, thresh)
}
\arguments{
\item{tseries}{(NumericVector) Time series of drought duration (time steps
not in drought have a duration of zero).  CAUTION:  This time series will be
modified.  See details.}

\item{thresh}{(double) Minimum drought duration.  Shorter droughts will
be eliminated.}
}
\value{
(NumericVector) Time series with droughts shorter than the threshold removed.
}
\description{
Eliminate any droughts with a duration less than the specified threshold.
}
\details{
Note that this function modifies the tseries argument.  This is done in order to
avoid the overhead of copying the vector.  Since our intended use is to \code{apply}
this to the rows of a matrix, this is of no consequence, since apply will do its own
copying.  However, if the function is used on a vector, users should be prepared for
this un-R-like behavior.
}
