% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{gents}
\alias{gents}
\title{Generate a matrix of drought time series for a selection of groups.}
\usage{
gents(ngroup, nmonth, Tg = rep(0, nmonth), expalpha = 1/300,
  beta = 0.1, L0 = 12, L1 = 24)
}
\arguments{
\item{ngroup}{Number of groups}

\item{nmonth}{Number of months}

\item{Tg}{Temperature}

\item{expalpha}{Baseline probability per month of a drought beginning}

\item{beta}{Temperature coefficient for relative risk}

\item{L0}{Minimum drought length}

\item{L1}{Maximum drought length}
}
\description{
A "group" is a geographical identifier used to tag observations from the same location,
whether it's a basin, a grid cell, or something else.
}
