% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weirdsci_demo.R
\name{get_ncdf_pixel}
\alias{get_ncdf_pixel}
\title{Pull temperature from a single grid cell in a group of netCDF files}
\usage{
get_ncdf_pixel(filelist, lat = 39.7, lon = -78.3, varnames = NULL)
}
\arguments{
\item{filelist}{Names of netCDF files to read.}

\item{lat}{Latitude of pixel to pull.}

\item{lon}{Longitude of pixel to pull.}

\item{varnames}{Named list of variable names}
}
\value{
Data frame with time, global temperature, and grid cell temperature.
}
\description{
Retrieve the temperature for a single grid cell specified by latitude and
longitude.
}
\details{
The varnames list allows you to override the default variable names.  You
only have to specify the ones that are changed from their defaults.  The
names recognized are:  latvar, lonvar, timevar, varname.
}
