% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnpy.R
\name{get_subblk}
\alias{get_subblk}
\title{Read a specific group of cells from a list of python pickle files}
\usage{
get_subblk(filelist, cellrng)
}
\arguments{
\item{filelist}{List of input filenames}

\item{cellrng}{Vector of cell indices to extract}
}
\value{
A list of matrices.  Each matrix contains the merged set of runs from all of
the input files for a single grid cell.
}
\description{
Given a list of python pickle files, each containing runs for the same block of
grid cells, extract the same subblock from each input file and merge them into a
single dataset with all of the runs for each grid cell.
}
\details{
The input files should be python pickle files, each containing a list of matrices.
Each matrix should contain runs for a single grid cell, with runs in rows and time
steps in columns.  All of the input files should contain data for the same grid cells,
in the same order.
}
