% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ticker}
\alias{plot_ticker}
\title{Make the ticker plot of drought incidence over time}
\usage{
plot_ticker(tsmat, thresh, yrstrt, col = "darkgrey", siz = 0.7)
}
\arguments{
\item{tsmat}{Matrix of time series: ensemble runs in rows, time (months) in
columns}

\item{thresh}{Duration threshold.  Droughts less than this duration will be
filtered.}

\item{yrstrt}{Start year.  Used to compute axis labels}

\item{col}{Color to use to plot the tickers}

\item{siz}{Size to plot the ticker dots.}
}
\description{
Make the ticker plot of drought incidence over time
}
