% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnpy.R
\name{readNPY}
\alias{readNPY}
\title{Read an array or matrix from a numpy save file}
\usage{
readNPY(filename)
}
\arguments{
\item{filename}{Name of the file to read}
}
\value{
An R array read from the data in the file.
}
\description{
Read an array or matrix from a numpy save file
}
