% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weirdsci_demo.R
\name{reformat_data}
\alias{reformat_data}
\title{Format data for the plots that we want to make}
\usage{
reformat_data(df)
}
\arguments{
\item{df}{Input data frame}
}
\value{
Data frame with colnames adjusted and extra columns added.
}
\description{
Input to this function is the data frame from \code{\link{get_ncdf_pixel}}.
}
