% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{gents}
\alias{gents}
\title{Generate a matrix of drought time series for a selection of basins.}
\usage{
gents(nbasin, nmonth, Tg = rep(0, nmonth), expalpha = 1/300,
  beta = 0.1, L0 = 12, L1 = 24)
}
\arguments{
\item{nbasin}{Number of basins}

\item{nmonth}{Number of months}

\item{Tg}{Temperature}

\item{expalpha}{Baseline probability per month of a drought beginning}

\item{beta}{Temperature coefficient for relative risk}

\item{L0}{Minimum drought length}

\item{L1}{Maximum drought length}
}
\description{
Generate a matrix of drought time series for a selection of basins.
}
