% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnpy.R
\name{readpkl}
\alias{readpkl}
\title{Read a pickled python object from a file}
\usage{
readpkl(filename)
}
\arguments{
\item{filename}{Name of the file to read}
}
\value{
An R object that more or less works like the python object; see details.
}
\description{
Read data saved with the python \code{pickle} module.
}
\details{
Python lists will return come through as lists; numpy arrays will come through as R
arrays (or matrices in 2-d).  I haven't tested what dictionaries turn into, but I'd
guess something like a named list.  If you try it sometime, let me know.
}
