% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{ts2event}
\alias{ts2event}
\alias{tsmat2event}
\title{Convert a single-basin drought time series into a data frame of event data}
\usage{
ts2event(ts, Tg, basinid, scenarioid, nbasin = 300, neventmax = 100,
  censored = FALSE)

tsmat2event(tsmat, Tg, scenarioid, nbasin = 300, neventmax = 100,
  censored = FALSE)
}
\arguments{
\item{ts}{Vector of monthly drought status.  0= no drought; >0= drought}

\item{Tg}{Vector of \emph{annual} global mean temperature.}

\item{basinid}{Numerical identifier for the basin.  This \emph{must} be
unique for each basin.}

\item{scenarioid}{Numerical identifier for the scenario.  This \emph{must} be
unique across scenarios.}

\item{nbasin}{Total number of basins.  This is used in constructing unique id
values from the scenarioid and basinid.}

\item{neventmax}{Maximum number of events for a single basin and scenario.
This is also used to construct UIDs.}

\item{censored}{Flag indicating whether we should account for left
and right censoring (see details).}

\item{tsmat}{A matrix of basin time series. Each row of the matrix should be
a time series for one basin.}
}
\value{
Data frame of event data.
}
\description{
The vector of status values will be converted into a data frame that
describes the waiting times for drought events and the changes in their
time-varying covariates.  Currently, the only covariate we are tracking is
Global mean temperature. Changes in Tg will be recorded at annual
boundaries.
}
\details{
An event data frame consists of the following columns:
\describe{
  \item{id}{A unique serial number for each drought event.}
  \item{basinid}{A unique id for each basin.  This id will be used as a
grouping variable in frailty models.}
  \item{tstart}{Start time for a recording period, in months since the last event (or
start of the scenario).  Covariate changes take effect at the start of a
period.}
  \item{tstop}{End time for a recording period, in months since the last
event.  Events during a period are deemed to occur at the end of the period.}
  \item{Tg}{Temperature covariate}
  \item{drought}{Indicator of whether the period had a drought.}
}

This processing prepares the data in a time series of drought/no drought
indicators to be used in a survival analysis.  Each basin can have multiple
events, and these will be treated as separate subjects.  The reason we do
this is that it's the easiest way to account for the fact that droughts have
duration, and while a drought is in progress you can't have another drought.
This turns out to be rather difficult to do in a counting formulation.

The intent is to use this data in a proportional hazard model, so by default
right censored events are not included.  Likewise, we make no effort to
adjust for the fact that for the first drought in a time series, we don't
actually know the full time since the last drought.
}
\section{Functions}{
\itemize{
\item \code{tsmat2event}: Convert a matrix of basin time series to an event data frame
}}

\section{Note}{

Setting the \code{censored} flag to \code{TRUE} is not currently supported.
}

